Create useChatCompletion hook to reuse playground code
